Modified the player.lua files for each city to lower faction if the race that inherits the city is player killed.
The event was changed from event_pvp_slay (custom event from RoT) to event_death (stock event) and the client pointers modified correctly. In addition a rule was added to code PVP:FactionCityKilling to enable/disable this feature, default is disabled.

Original files is credited to Rise of Thule (Ghansta) box for implementing this feature.